+++ R/livy_sources.R
@@ -6,7 +6,21 @@ livy_sources_included <- function() {
     "/stream\\.scala",
     "/sqlutils\\.scala",
     "/utils\\.scala",
-    "/repartition\\.scala"
+    "/repartition\\.scala",
+    "/tracker\\.scala",
+    "/livyutils\\.scala",
+    "/applyutils\\.scala",
+    "/classutils\\.scala",
+    "/fileutils\\.scala",
+    "/rscript\\.scala",
+    "/sources\\.scala",
+    "/workercontext\\.scala",
+    "/channel\\.scala",
+    "/handler\\.scala",
+    "/backend\\.scala",
+    "/workerhelper\\.scala",
+    "/workerrdd\\.scala",
+    "/workerutils\\.scala"
   )
 }
 
@@ -29,6 +43,9 @@ livy_sources_refresh <- function() {
                                  basename(sourcePath))
 
     lines <- readLines(sourcePath)
+
+    if (!dir.exists(dirname(destinationPath)))
+      dir.create(dirname(destinationPath), recursive = TRUE)
 
     write("//", file = destinationPath, append = FALSE)
     write("// This file was automatically generated using livy_sources_refresh()", file = destinationPath, append = TRUE)