+++ R/spark_compile_embedded.R
@@ -1,17 +1,15 @@
 spark_compile_embedded_sources <- function() {
   worker_files <- dir("R", full.names = TRUE, pattern = "worker|core")
   rlines <- unlist(lapply(worker_files, function(e) readLines(e)))
-  rlines <- gsub("\\\\", "\\\\\\\\", rlines)
-  rlines <- gsub("\\\"", "\\\\\"", rlines)
   rlines <- c(rlines, "do.call(spark_worker_main, as.list(commandArgs(trailingOnly = TRUE)))")
 
   lines <- c(
     "package sparklyr",
     "",
-    "object Sources {",
-    "  def sources: String = \"\" +",
-    paste("    \"", rlines, "\\n\" +", sep = ""),
-    "    \"\"",
+    "class Sources {",
+    "  def sources: String = \"\"\"",
+    rlines,
+    "    \"\"\"",
     "}"
   )