+++ java/spark-1.5.2/stream.scala
@@ -1,16 +1,16 @@
 package sparklyr
 
-import java.io.{ByteArrayInputStream, ByteArrayOutputStream, DataInputStream, DataOutputStream}
+class StreamHandler(serializer: Serializer, tracker: JVMObjectTracker) {
 
-import scala.collection.mutable.HashMap
-import scala.language.existentials
+  import java.io._
 
-import io.netty.channel.{ChannelHandlerContext, SimpleChannelInboundHandler}
-import io.netty.channel.ChannelHandler.Sharable
+  import scala.collection.mutable.HashMap
+  import scala.language.existentials
 
-import sparklyr.Serializer._
+  import io.netty.channel.{ChannelHandlerContext, SimpleChannelInboundHandler}
+  import io.netty.channel.ChannelHandler.Sharable
 
-object StreamHandler {
+  val invoke = new Invoke()
 
   def read(
     msg: Array[Byte],
@@ -24,10 +24,10 @@ object StreamHandler {
     val bos = new ByteArrayOutputStream()
     val dos = new DataOutputStream(bos)
 
-    val objId = readString(dis)
-    val isStatic = readBoolean(dis)
-    val methodName = readString(dis)
-    val numArgs = readInt(dis)
+    val objId = serializer.readString(dis)
+    val isStatic = serializer.readBoolean(dis)
+    val methodName = serializer.readString(dis)
+    val numArgs = serializer.readInt(dis)
 
     if (objId == "Handler") {
       methodName match {
@@ -35,34 +35,49 @@ object StreamHandler {
           val args = readArgs(numArgs, dis)
           if (numArgs != 1) throw new IllegalArgumentException("echo should take a single argument")
 
-          writeInt(dos, 0)
-          writeObject(dos, args(0))
+          serializer.writeInt(dos, 0)
+          serializer.writeObject(dos, args(0))
         case "rm" =>
           try {
-            val t = readObjectType(dis)
+            val t = serializer.readObjectType(dis)
             if (t != 'c') throw new IllegalArgumentException("object removal expects a string")
-            val objToRemove = readString(dis)
-            JVMObjectTracker.remove(objToRemove)
-            writeInt(dos, 0)
-            writeObject(dos, null)
+            val objToRemove = serializer.readString(dis)
+            tracker.remove(objToRemove)
+            serializer.writeInt(dos, 0)
+            serializer.writeObject(dos, null)
           } catch {
             case e: Exception =>
               logger.logError(s"failed to remove $objId", e)
-              writeInt(dos, -1)
-              writeString(dos, s"Removing $objId failed: ${e.getMessage}")
+              serializer.writeInt(dos, -1)
+              serializer.writeString(dos, s"Removing $objId failed: ${e.getMessage}")
           }
         case "getHostContext" =>
-          writeInt(dos, 0)
-          writeObject(dos, hostContext.asInstanceOf[AnyRef])
+          serializer.writeInt(dos, 0)
+          serializer.writeObject(dos, hostContext.asInstanceOf[AnyRef])
         case _ =>
           dos.writeInt(-1)
-          writeString(dos, s"Error: unknown method $methodName")
+          serializer.writeString(dos, s"Error: unknown method $methodName")
       }
     } else {
       handleMethodCall(isStatic, objId, methodName, numArgs, dis, dos, classMap, logger)
     }
 
     bos.toByteArray
+  }
+
+  /**
+   * Return a nice string representation of the exception. It will call "printStackTrace" to
+   * recursively generate the stack trace including the exception and its causes.
+   */
+  def exceptionString(e: Throwable): String = {
+    if (e == null) {
+      "No exception information provided."
+    } else {
+      // Use e.printStackTrace here because e.getStackTrace doesn't include the cause
+      val stringWriter = new StringWriter()
+      e.printStackTrace(new PrintWriter(stringWriter))
+      stringWriter.toString
+    }
   }
 
   def handleMethodCall(
@@ -85,7 +100,7 @@ object StreamHandler {
             Class.forName(objId)
           }
         } else {
-          JVMObjectTracker.get(objId) match {
+          tracker.get(objId) match {
             case None => throw new IllegalArgumentException("Object not found " + objId)
             case Some(o) =>
               obj = o
@@ -94,21 +109,21 @@ object StreamHandler {
         }
 
         val args = readArgs(numArgs, dis)
-        val res = Invoke.invoke(cls, objId, obj, methodName, args, logger)
+        val res = invoke.invoke(cls, objId, obj, methodName, args, logger)
 
-        writeInt(dos, 0)
-        writeObject(dos, res.asInstanceOf[AnyRef])
+        serializer.writeInt(dos, 0)
+        serializer.writeObject(dos, res.asInstanceOf[AnyRef])
       } catch {
         case e: Exception =>
           logger.logError(s"failed calling $methodName on $objId")
-          writeInt(dos, -1)
-          writeString(dos, Utils.exceptionString(
+          serializer.writeInt(dos, -1)
+          serializer.writeString(dos, exceptionString(
             if (e.getCause == null) e else e.getCause
           ))
         case e: NoClassDefFoundError =>
           logger.logError(s"failed calling $methodName on $objId with no class dound error")
-          writeInt(dos, -1)
-          writeString(dos, Utils.exceptionString(
+          serializer.writeInt(dos, -1)
+          serializer.writeString(dos, exceptionString(
             if (e.getCause == null) e else e.getCause
           ))
       }
@@ -117,7 +132,7 @@ object StreamHandler {
   // Read a number of arguments from the data input stream
   def readArgs(numArgs: Int, dis: DataInputStream): Array[java.lang.Object] = {
     (0 until numArgs).map { _ =>
-      readObject(dis)
+      serializer.readObject(dis)
     }.toArray
   }
 }