+++ java/spark-1.6.0/workercontext.scala
@@ -1,12 +1,7 @@
 package sparklyr
 
-import org.apache.spark._
-import org.apache.spark.rdd.RDD
-import org.apache.spark.sql._
-import scala.collection.JavaConversions._
-
 class WorkerContext(
-  sourceArray: Array[Row],
+  sourceArray: Array[org.apache.spark.sql.Row],
   lock: AnyRef,
   closure: Array[Byte],
   columns: Array[String],
@@ -14,6 +9,11 @@ class WorkerContext(
   closureRLang: Array[Byte],
   bundlePath: String,
   context: Array[Byte]) {
+
+  import org.apache.spark._
+  import org.apache.spark.rdd.RDD
+  import org.apache.spark.sql._
+  import scala.collection.JavaConversions._
 
   private var result: Array[Row] = Array[Row]()
 
@@ -69,51 +69,5 @@ class WorkerContext(
 
   def getContext(): Array[Byte] = {
     context
-  }
-}
-
-object WorkerHelper {
-  def computeRdd(
-    rdd: RDD[Row],
-    closure: Array[Byte],
-    config: String,
-    port: Int,
-    columns: Array[String],
-    groupBy: Array[String],
-    closureRLang: Array[Byte],
-    bundlePath: String,
-    customEnv: java.util.Map[Object, Object],
-    connectionTimeout: Int,
-    context: Array[Byte],
-    options: java.util.Map[Object, Object]
-  ): RDD[Row] = {
-
-    var customEnvMap = scala.collection.mutable.Map[String, String]();
-    customEnv.foreach(kv => customEnvMap.put(
-      kv._1.asInstanceOf[String],
-      kv._2.asInstanceOf[String])
-    )
-
-    var optionsMap = scala.collection.mutable.Map[String, String]();
-    options.foreach(kv => optionsMap.put(
-      kv._1.asInstanceOf[String],
-      kv._2.asInstanceOf[String])
-    )
-
-    val computed: RDD[Row] = new WorkerRDD(
-      rdd,
-      closure,
-      columns,
-      config,
-      port,
-      groupBy,
-      closureRLang,
-      bundlePath,
-      Map() ++ customEnvMap,
-      connectionTimeout,
-      context,
-      Map() ++ optionsMap)
-
-    computed
   }
 }